# Optimization batch 6: make full use of exact renames